feat: add getting user by id, divide endpoints into separated functions
