+++ src/types.ts
@@ -1,3 +1,5 @@
+import { IncomingMessage, ServerResponse } from 'http';
+
 export type User = {
   id: string;
   username: string;
@@ -6,3 +8,7 @@ export type User = {
 };
 
 export type BodyUser = Omit<User, 'id'>;
+
+export type ReqType = IncomingMessage;
+
+export type ResType = ServerResponse<ReqType> & { req: ReqType };
